Changed clock field type to BigInteger (issue migasfree/migasfree#241)

+++ migasfree/hardware/models/node.py
@@ -173,7 +173,7 @@ class Node(models.Model):
         null=True
     )
 
-    clock = models.IntegerField(
+    clock = models.BigIntegerField(
         verbose_name=_("clock"),
         null=True
     )
